+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/dalance/nom-tracable/compare/v0.2.1...Unreleased) - ReleaseDate
 
+* [Added] custom field to TracableInfo
+
 ## [v0.2.1](https://github.com/dalance/nom-tracable/compare/v0.2.0...v0.2.1) - 2019-07-29
 
 * [Added] custom_trace function