+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased](https://github.com/dalance/nom-tracable/compare/v0.2.1...Unreleased) - ReleaseDate
 
 * [Added] custom field to TracableInfo
+* [Added] fold support
 
 ## [v0.2.1](https://github.com/dalance/nom-tracable/compare/v0.2.0...v0.2.1) - 2019-07-29